Fix VAE precisions (#588)

+++ fastvideo/v1/configs/pipelines/wan.py
@@ -50,7 +50,7 @@ class WanT2V480PConfig(PipelineConfig):
 
     # Precision for each component
     precision: str = "bf16"
-    vae_precision: str = "fp16"
+    vae_precision: str = "fp32"
     text_encoder_precisions: Tuple[str, ...] = field(
         default_factory=lambda: ("fp32", ))
